Add GCC 14.2.0 to COMPILERS.md
So in exec README it is mentioned that it is confirmed to work with GCC 14.2.0. I assumed that whole OS has been compiled with it, so it's a reason to add 14.2.0 to supported GCC versions. Also, given the changes in code, I think I'll have to re-verify if 13.3.0 still works.